requirements.txt
psycopg2-binary
faker
mysql-connector-python
pandas
configparser



